+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.3
+
+Created an example and wrote better single API documentation.
+
+## 0.0.2
+
+Added better documentation and examples for the panorama_creator package.
+
 ## 0.0.1
 
 Initial release of the panorama_creator package.